refactor: refactoring folders and files of constants
