+++ src/constants/places/farmacias.js
