--- src/constants/trips.js
@@ -1,25 +1,0 @@
-import abaetetuba4bocas from './trips/abaetetuba-4bocas'
-import abaetetubaBarcarena from './trips/abaetetuba-barcarena'
-import barcarenaBelem from './trips/barcarena-belem'
-import abaetetubaBelem from './trips/abaetetuba-belem'
-import abaetetubaIgarapemiri from './trips/abaetetuba-igarapemiri'
-import abaetetubaMaedorio from './trips/abaetetuba-maedorio'
-import abaetetubaMoju from './trips/abaetetuba-moju'
-import abaetetubaNucleoPortoSaoFrancisco from './trips/abaetetuba-nucleo-portoSaoFrancisco'
-import abaetetubaPortoSaoFrancisco from './trips/abaetetuba-portoSaoFrancisco'
-import abaetetubaBeja from './trips/abaetetuba-beja'
-import abaetetubaUfpa from './trips/abaetetuba-ufpa'
-
-export const trips = [
-  { ...abaetetubaBelem },
-  { ...abaetetubaUfpa },
-  { ...abaetetubaBarcarena },
-  { ...barcarenaBelem },
-  { ...abaetetubaBeja },
-  { ...abaetetubaMoju },
-  { ...abaetetubaIgarapemiri },
-  { ...abaetetubaPortoSaoFrancisco },
-  { ...abaetetubaNucleoPortoSaoFrancisco },
-  { ...abaetetubaMaedorio },
-  { ...abaetetuba4bocas }
-]
