+++ src/constants/trips/index.js
@@ -1,0 +1,25 @@
+import abaetetuba4bocas from './abaetetuba-4bocas'
+import abaetetubaBarcarena from './abaetetuba-barcarena'
+import barcarenaBelem from './barcarena-belem'
+import abaetetubaBelem from './abaetetuba-belem'
+import abaetetubaIgarapemiri from './abaetetuba-igarapemiri'
+import abaetetubaMaedorio from './abaetetuba-maedorio'
+import abaetetubaMoju from './abaetetuba-moju'
+import abaetetubaNucleoPortoSaoFrancisco from './abaetetuba-nucleo-portoSaoFrancisco'
+import abaetetubaPortoSaoFrancisco from './abaetetuba-portoSaoFrancisco'
+import abaetetubaBeja from './abaetetuba-beja'
+import abaetetubaUfpa from './abaetetuba-ufpa'
+
+export const trips = [
+  { ...abaetetubaBelem },
+  { ...abaetetubaUfpa },
+  { ...abaetetubaBarcarena },
+  { ...barcarenaBelem },
+  { ...abaetetubaBeja },
+  { ...abaetetubaMoju },
+  { ...abaetetubaIgarapemiri },
+  { ...abaetetubaPortoSaoFrancisco },
+  { ...abaetetubaNucleoPortoSaoFrancisco },
+  { ...abaetetubaMaedorio },
+  { ...abaetetuba4bocas }
+]
