+++ src/pages/MapCity.vue
@@ -88,7 +88,7 @@
 
 <script>
 import L from 'leaflet'
-import { makers } from 'src/constants/makersMap'
+import { makers } from 'src/constants/places/index'
 export default {
   name: 'MapPage',
   data () {
@@ -117,9 +117,24 @@ export default {
           icon: 'flat/acai.png'
         },
         {
+          label: 'Artigos de Pesca Esportiva',
+          value: 'Pesca',
+          icon: 'flat/fishing.png'
+        },
+        {
           label: 'Bancos',
           value: 'Bancos',
           icon: 'flat/bank.png'
+        },
+        {
+          label: 'Churrascarias/Espetarias',
+          value: 'Churrascaria&Espetaria',
+          icon: 'categories/churrascaria.png'
+        },
+        {
+          label: 'Farmácias',
+          value: 'Farmácias',
+          icon: 'places/farmacia.png'
         },
         {
           label: 'Hotéis',
@@ -132,14 +147,9 @@ export default {
           icon: 'places/hospital.png'
         },
         {
-          label: 'Farmácias',
-          value: 'Farmácias',
-          icon: 'places/farmacia.png'
-        },
-        {
-          label: 'Churrascarias/Espetarias',
-          value: 'Churrascaria&Espetaria',
-          icon: 'categories/churrascaria.png'
+          label: 'Igrejas',
+          value: 'Igrejas',
+          icon: 'flat/church.png'
         },
         {
           label: 'Pub & Restaurante',
@@ -151,11 +161,7 @@ export default {
           value: 'Sorveterias',
           icon: 'categories/sorveteria.png'
         },
-        {
-          label: 'Artigos de Pesca Esportiva',
-          value: 'Pesca',
-          icon: 'flat/fishing.png'
-        },
+
         {
           label: 'Suplementos e Saúde',
           value: 'Suplemento&Saude',
@@ -237,3 +243,5 @@ export default {
   /* min-height: 90vh; */
 }
 </style>
+src/constants/places/makersMap
+src/constants/places
